+++ ydb/services/sqs_topic/queue_url/ut/ya.make
@@ -1,0 +1,15 @@
+GTEST()
+
+SRCS(
+    utils_ut.cpp
+)
+
+PEERDIR(
+    ydb/services/sqs_topic/queue_url
+)
+
+SRCDIR(
+    ydb/services/sqs_topic/queue_url
+)
+
+END()
