+++ ydb/services/sqs_topic/queue_url/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    utils.cpp
+)
+
+PEERDIR(
+    library/cpp/string_utils/url
+)
+
+END()
+
+RECURSE_FOR_TESTS(ut)
