+++ ydb/services/sqs_topic/utils.h
@@ -1,17 +1,3 @@
 #pragma once
 
-#include <util/generic/string.h>
-
-namespace NKikimr::NSqsTopic {
-
-    struct TRichQueueUrl {
-        TString Database;
-        TString Consumer;
-        TString TopicPath;
-        bool Fifo{};
-    };
-
-    TRichQueueUrl ParseQueueUrl(const TStringBuf queueUrl);
-
-    TString PackQueueUrlPath(const TRichQueueUrl& queueUrl);
-}
+#include <ydb/services/sqs_topic/queue_url/utils.h>
