+++ ydb/services/sqs_topic/ya.make
@@ -19,9 +19,11 @@ PEERDIR(
     ydb/services/lib/actors
     ydb/services/lib/sharding
     ydb/services/persqueue_v1
+    ydb/services/sqs_topic/queue_url
     ydb/services/ydb
     ydb/services/ymq
-    library/cpp/string_utils/url
 )
 
 END()
+
+RECURSE(queue_url)
